% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Candidate.R
\docType{class}
\name{Candidate-class}
\alias{Candidate-class}
\alias{initialize,Candidate-method}
\alias{show,Candidate-method}
\title{A Candidate object}
\description{
Object of class \code{Candidate} are created by the \code{CreateCandidate} functionsn. Objects of
this class include \code{show} and \code{print} methods.
}
\details{
An object of the class `Candidate' has the following slots:
\itemize{
\item \code{name} The name of the candidate
\item \code{Party} The name of the party
\item \code{delegatesWon} Number of delegates won
\item \code{delegatesNeeded} Number of delegates needed to win nomination
}
}
\author{
Mauricio Vela
}

