% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createCandidate.R
\name{createCandidate}
\alias{createCandidate}
\alias{createCandidate,ANY-method}
\title{Create an object of class 'Candidate'
Function that creates an object of class `Candidate'}
\usage{
createCandidate(name, delegatesWon, party)
}
\arguments{
\item{name}{Name of the presidential candidate}

\item{delegatesWon}{Number of delegates won by given candidate}

\item{party}{Democrat or Republican}
}
\value{
An object of class `Candidate' containing
\itemize{
\item \code{name} Name of the presidential candidate
\item \code{party} Party 
\item \code{delegatesWon} Delegates won
\item \code{delegatesNeeded} Number of delegates needed to obtain a nomination
 }
}
\description{
Create an object of class 'Candidate'
Function that creates an object of class `Candidate'
}
\examples{

Candidate <-createCandidate("Obama", 100, "Democrat") 

}
\author{
Mauricio Vela
}
\seealso{
\code{\link{PropNeeded}}, \code{\link{createRace}}
}

