% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createRace.R
\name{createRace}
\alias{createRace}
\alias{createRace,ANY-method}
\title{Create an object of class 'Race'}
\usage{
createRace(Candidates, party)
}
\arguments{
\item{party}{Party}

\item{name}{List of canidates of object class 'Candidate'}
}
\value{
An object of class 'Race' containing
\itemize{
\item Candidates  of class 'Candidate'
\item party Party 
\item delegatesNeeded Number of delegates needed to win
\item delegatesRemaining Number of delegates remaining
}
}
\description{
Function creates an object of class 'Race'
}
\examples{
clinton <- createCandidate("Clinton", 1614, "Democrat")
sanders <- createCandidate("Sanders", 856, "Democrat")
Race <- createRace(Candidates = list(clinton, sanders), party = "Democrat")
plot(Race)

}
\author{
Mauricio Vela
}

