\name{primarypackage-package}
\alias{primarypackage-package}
\alias{primarypackage}
\docType{package}
\title{
\packageTitle{primarypackage}
}
\description{
\packageDescription{primarypackage}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{primarypackage}
\packageIndices{primarypackage}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
\packageAuthor{primarypackage}

Maintainer: \packageMaintainer{primarypackage}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in the R documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
