% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/primarypackage-package.R
\docType{package}
\name{primarypackage}
\alias{primarypackage}
\alias{primarypackage-package}
\title{primarypackage}
\description{
The primarypackage package creates candidate and calculate delegates
}
\author{
Mauricio Vela: \email{mauriciovelabaron@wustl.edu}
}
\seealso{
\code{\link{createCandidate}}, \code{\link{Candidate}}, \code{\link{propNeeded}}
}

