% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Race.R
\docType{class}
\name{Race-class}
\alias{Race-class}
\alias{initialize,Race-method}
\alias{plot,Race-method}
\title{A Race object 
Objects of class \code{Race} are created by the \code{createRace} function.}
\description{
An object of the class `Race' has the following slots:
\itemize{
\item \code{Candidates} Objects of class 'Candidate' from the same party
\item \code{party} Party 
\item \code{delegatesNeeded} Number of delegates needed
\item \code{delegatesRemaining} Number of delegates remaining
}
}
\examples{
clinton <- createCandidate("Clinton", 1614, "Democrat")
sanders <- createCandidate("Sanders", 856, "Democrat")
Race <- createRace(Candidates = list(clinton, sanders), party = "Democrat")
plot(Race)

}
\author{
Mauricio Vela
}
\seealso{
\code{\link{createRace}}
}

